move user info logic off command, share with api
